Configurado timestamp arquivo csv

+++ index.js
@@ -8,15 +8,20 @@ const config = require('./config.json');
 async function execAll() {
   logger.info('Start -> [all]');
   console.time('all');
-  logger.info('DOWNLOAD');
-  await sftp.downloadAll();
-  logger.info('CONVERT');
-  xlsCsv.convertAll();
-  logger.info('UPLOAD');
-  await sftp.uploadAll();
-  logger.info('DELETE');
-  utils.deleteAll(`${config.xlsfolder}`);
-  utils.deleteAll(`${config.csvfolder}`);
+  let files = await sftp.listFiles();
+  if (files.length > 0){
+    logger.info('DOWNLOAD');
+    await sftp.downloadAll();
+    logger.info('CONVERT');
+    xlsCsv.convertAll();
+    logger.info('UPLOAD');
+    await sftp.uploadAll();
+    logger.info('DELETE');
+    utils.deleteAll(`${config.xlsfolder}`);
+    utils.deleteAll(`${config.csvfolder}`);
+  } else {
+    logger.error('Files not found');
+  }
   console.timeEnd('all');
   logger.info('End -> [all]');
 }
@@ -48,15 +53,20 @@ async function execOne() {
 async function execFolder() {
   logger.info('Start -> [folder]');
   console.time('folder');
-  logger.info('DOWNLOAD');
-  await sftp.downloadFolder();
-  logger.info('CONVERT');
-  xlsCsv.convertAll();
-  logger.info('UPLOAD');
-  await sftp.uploadFolder();
-  logger.info('DELETE');
-  utils.deleteAll(`${config.xlsfolder}`);
-  utils.deleteAll(`${config.csvfolder}`);
+  let files = await sftp.listFiles();
+  if (files.length > 0){
+    logger.info('DOWNLOAD');
+    await sftp.downloadFolder();
+    logger.info('CONVERT');
+    xlsCsv.convertAll();
+    logger.info('UPLOAD');
+    await sftp.uploadFolder();
+    logger.info('DELETE');
+    utils.deleteAll(`${config.xlsfolder}`);
+    utils.deleteAll(`${config.csvfolder}`);
+  } else {
+    logger.error('Files not found');
+  }
   console.timeEnd('folder');
   logger.info('End -> [folder]');
 }
